Address final code review feedback
- Add z.number().int() validation for processCount to match Go int type
- Remove hardcoded GOARCH from Dockerfile to support multi-arch builds
- Docker BuildKit automatically sets TARGETARCH for the target platform

+++ packages/api/src/routers/metrics.ts
@@ -15,7 +15,7 @@ const systemMetricsSchema = z.object({
     loadAverage1: z.number(),
     loadAverage5: z.number(),
     loadAverage15: z.number(),
-    processCount: z.number(),
+    processCount: z.number().int(),
 });
 
 const containerMetricsSchema = z.object({
